adcionado tratamento de exceções

+++ src/main/java/com/example/meu_primeiro_springboot/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package com.example.meu_primeiro_springboot.service;
 
+import com.example.meu_primeiro_springboot.exceptions.RecursoNaoEncontradoException;
 import com.example.meu_primeiro_springboot.model.Produto;
 import com.example.meu_primeiro_springboot.repository.ProdutoRepository;
 import org.springframework.stereotype.Service;
@@ -19,11 +20,15 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarPorId(Long id){
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(Long id){
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto do id: "+ id +" nao encontrado"));
     }
 
     public void delete(Long id){
+        if(!produtoRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Produto com id "+ id +" nao encontrado");
+        }
         produtoRepository.deleteById(id);
     }
